+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - `collection set description` sub command
 - `op set description` sub command
 - `op create` sub command
+- made `opctl run` autopopulate op params from env
 
 ### Removed
 - `add-op` sub command